requirements.txt
requests
python-dotenv
google-generativeai
beautifulsoup4
pandas
sqlalchemy
streamlit